ai-interact/requirements.txt
streamlit
google-genai
python-dotenv
mcp-server-fetch
mcp==1.9.0